Revert "Merge branch 'main' of github.com:adobe/react-spectrum into storybook-10-upgrade"
This reverts commit bf62a86810b8e677fdd8c609bc923f9339e19213, reversing
changes made to edc24b0ff9fdc9639fc062deb181f8ffc52c5331.
