+++ .chromatic/main.js
@@ -1,5 +1,9 @@
 
-export default {
+module.exports = {
+  framework: {
+    name: "storybook-react-parcel",
+    options: {},
+  },
   stories: [
     '../packages/**/chromatic/**/*.stories.@(js|jsx|ts|tsx)',
     '../packages/@react-spectrum/s2/chromatic/*.stories.@(js|jsx|mjs|ts|tsx)'
@@ -8,7 +12,8 @@ export default {
     'storybook/actions',
     '@storybook/addon-a11y'
   ],
-  typescript: { check: false, reactDocgen: false },
-  framework: { name: 'storybook-react-parcel', options: {} },
-  core: { disableWhatsNewNotifications: true },
+  typescript: {
+    check: false,
+    reactDocgen: false
+  }
 };
