+++ .storybook-s2/custom-addons/provider/register.js
@@ -4,9 +4,8 @@ import {locales} from '../../constants';
 import React, {useEffect, useState} from 'react';
 
 function ProviderFieldSetter({api}) {
-  let [values, setValues] = useState(() => ({
-    locale: api.getQueryParam('providerSwitcher-locale') || undefined,
-  }));
+  let localeParam = api.getQueryParam('providerSwitcher-locale') || undefined;
+  let [values, setValues] = useState({locale: localeParam});
   let channel = addons.getChannel();
   let onLocaleChange = (e) => {
     let newValue = e.target.value || undefined;
@@ -24,19 +23,19 @@ function ProviderFieldSetter({api}) {
     return () => {
       channel.removeListener('rsp/ready-for-update', storySwapped);
     };
-  }, [values, channel]);
+  });
 
   useEffect(() => {
     api.setQueryParams({
       'providerSwitcher-locale': values.locale || ''
     });
-  }, [api, values.locale]);
+  });
 
   return (
     <div style={{display: 'flex', alignItems: 'center', fontSize: '12px'}}>
       <div style={{marginRight: '10px'}}>
         <label htmlFor="locale">Locale: </label>
-        <select id="locale" name="locale" onChange={onLocaleChange} value={values.locale || ''}>
+        <select id="locale" name="locale" onChange={onLocaleChange} value={values.locale}>
           {locales.map(locale => <option key={locale.label} value={locale.value}>{locale.label}</option>)}
         </select>
       </div>
