+++ .storybook/main.js
@@ -1,8 +1,5 @@
-import { fileURLToPath } from "node:url";
 
-const localAddon = (rel) => fileURLToPath(import.meta.resolve(rel));
-
-export default {
+module.exports = {
   stories: [
     '../packages/@{react-aria,react-stately,spectrum-icons}/*/stories/*.stories.{js,jsx,ts,tsx}',
     '../packages/@react-spectrum/!(s2)/stories/*.stories.{js,jsx,ts,tsx}',
@@ -11,17 +8,29 @@ export default {
     '../packages/react-stately/stories/*/*.stories.{js,jsx,ts,tsx}',
     '../packages/react-aria-components/stories/*.stories.{js,jsx,ts,tsx}'
   ],
+
   addons: [
     'storybook/actions',
     '@storybook/addon-a11y',
     '@vueless/storybook-dark-mode',
-    localAddon('./custom-addons/provider'),
-    localAddon('./custom-addons/descriptions'),
-    localAddon('./custom-addons/theme'),
-    localAddon('./custom-addons/strictmode'),
-    localAddon('./custom-addons/scrolling'),
+    './custom-addons/provider/register.js',
+    './custom-addons/descriptions/register.js',
+    './custom-addons/theme/register.js',
+    './custom-addons/strictmode/register.js',
+    './custom-addons/scrolling/register.js'
   ],
-  typescript: { check: false, reactDocgen: false },
-  framework: { name: 'storybook-react-parcel', options: {} },
-  core: { disableWhatsNewNotifications: true },
+
+  typescript: {
+    check: false,
+    reactDocgen: false
+  },
+
+  framework: {
+    name: 'storybook-react-parcel',
+    options: {}
+  },
+
+  core: {
+    disableWhatsNewNotifications: true
+  }
 };
