+++ package.json
@@ -119,12 +119,12 @@
     "@react-spectrum/s2-icon-builder": "workspace:^",
     "@spectrum-css/component-builder": "workspace:^",
     "@spectrum-css/vars": "^2.3.0",
-    "@storybook/addon-a11y": "^10.0.0",
-    "@storybook/addon-docs": "^10.0.0",
-    "@storybook/addon-jest": "10.0.0-beta.12",
-    "@storybook/addon-themes": "^10.0.0",
-    "@storybook/react": "^10.0.0",
-    "@storybook/test-runner": "^0.24.0",
+    "@storybook/addon-a11y": "^9.0.18",
+    "@storybook/addon-docs": "^9.0.18",
+    "@storybook/addon-jest": "^9.0.18",
+    "@storybook/addon-themes": "^9.0.18",
+    "@storybook/react": "^9.0.18",
+    "@storybook/test-runner": "^0.22.0",
     "@stylistic/eslint-plugin-ts": "^2.9.0",
     "@swc/core": "^1.3.36",
     "@swc/jest": "^0.2.36",
@@ -139,7 +139,7 @@
     "@vitejs/plugin-react": "^5.1.4",
     "@vitest/browser-playwright": "^4.0.17",
     "@vitest/browser-preview": "^4.0.17",
-    "@vueless/storybook-dark-mode": "^10.0.0",
+    "@vueless/storybook-dark-mode": "^9.0.6",
     "@yarnpkg/types": "^4.0.0",
     "autoprefixer": "^9.6.0",
     "axe-playwright": "^1.1.11",
@@ -199,7 +199,7 @@
     "rimraf": "^6.0.1",
     "shadow-dom-testing-library": "^1.13.1",
     "sharp": "^0.33.5",
-    "storybook": "^10.0.0",
+    "storybook": "^9.0.18",
     "storybook-react-parcel": "workspace:^",
     "tailwind-variants": "patch:tailwind-variants@npm%3A0.3.1#~/.yarn/patches/tailwind-variants-npm-0.3.1-48888516de.patch",
     "tailwindcss": "^4.0.0",
