+++ packages/dev/parcel-resolver-storybook/StorybookResolver.ts
@@ -1,32 +1,4 @@
-// Parcel resolver implementing the upstream "externalize the runtime" pattern that
-// both @storybook/builder-vite and @storybook/builder-webpack5 use.
-//
-// Upstream references:
-//   - code/builders/builder-vite/src/plugins/storybook-external-globals-plugin.ts
-//   - code/builders/builder-webpack5/src/preview/iframe-webpack.config.ts
-//     (`const externals: Record<string, string> = globalsNameReferenceMap`)
-//   - node_modules/storybook/dist/preview/globals.js (source of `globalsNameReferenceMap`)
-//
-// The map is read at resolver startup (not hardcoded) so any future Storybook
-// version that adds entries is picked up automatically. For each (specifier ->
-// __STORYBOOK_MODULE_*__) pair we emit one cache file under
-// `node_modules/.cache/sb-parcel-externals/<key>.js` -- mirroring Vite's
-// `node_modules/.cache/sb-vite-plugin-externals/` layout -- so the synthetic
-// module is a real on-disk file Parcel's graph can hash and watch.
-//
-// Why this matters: without externalization, Parcel's CJS dev bundle ends up
-// with `storybook/preview-api` <-> chunk-SZQXB3JV <-> `storybook/internal/csf`
-// in a real cycle. csf/index.js (line ~916) reads `addons` at top level, but
-// `addons` is still undefined when csf evaluates inside the cycle, throwing
-// `TypeError: Cannot read properties of undefined (reading 'addons')` during
-// Storybook init. See plan docs/superpowers/plans/2026-04-25-parcel-builder-
-// upstream-alignment.md "Starting Context" for the full diagnosis. Routing
-// every preview-runtime specifier to a leaf module that just reads from
-// globalThis breaks the back-edge and removes the cycle entirely.
-
 import path from 'path';
-import fs from 'fs';
-import { createRequire } from 'module';
 import { Resolver } from "@parcel/plugin";
 const reactVersion = require("react-dom/package.json").version;
 import { default as NodeResolver } from "@parcel/node-resolver-core";
@@ -35,134 +7,9 @@ import { isGlob, glob, normalizeSeparators, relativePath } from '@parcel/utils';
 
 const REACT_MAJOR_VERSION = parseInt(reactVersion.split('.')[0], 10);
 
-// Approach B = enumerate named exports at codegen and emit
-// `const { x, y } = globalThis.__STORYBOOK_MODULE_*__; export { x, y };`.
-// Approach A fallback = `module.exports = globalThis.__STORYBOOK_MODULE_*__;`,
-// used only when require() can't enumerate the source module's exports.
-
-const _require = createRequire(__filename);
-
-interface ExternalEntry {
-  filePath: string;
-  // Whether we used the Approach B (named re-export) or Approach A (CJS interop) fallback
-  approach: 'B' | 'A';
-}
-
-function buildExternalsCache(): Record<string, ExternalEntry> {
-  const result: Record<string, ExternalEntry> = {};
-  const cacheDir = path.join(
-    _require.resolve('storybook/package.json'),
-    '..', '..', '.cache', 'sb-parcel-externals'
-  );
-  fs.mkdirSync(cacheDir, { recursive: true });
-
-  // Read storybook's canonical map at startup; do not hardcode.
-  let globalsNameReferenceMap: Record<string, string>;
-  try {
-    const globalsModulePath = _require.resolve('storybook/internal/preview/globals');
-    // Use dynamic import via createRequire-style; storybook ships ESM here so
-    // fall back to reading the JS source if require() can't load ESM.
-    try {
-      const mod = _require(globalsModulePath);
-      globalsNameReferenceMap = mod.globalsNameReferenceMap;
-    } catch {
-      const src = fs.readFileSync(globalsModulePath, 'utf8');
-      // Match the object literal between `var globalsNameReferenceMap = {` and `}`
-      const m = src.match(/globalsNameReferenceMap\s*=\s*(\{[\s\S]*?\})\s*,/);
-      if (!m) throw new Error('Could not parse globalsNameReferenceMap from ' + globalsModulePath);
-      // Parse it as JSON-ish (keys are quoted strings, values are quoted strings,
-      // but there's a trailing comment line). Strip line comments first.
-      const cleaned = m[1].replace(/\/\/.*$/gm, '').replace(/,\s*\}/, '}');
-      // eslint-disable-next-line no-new-func
-      globalsNameReferenceMap = (new Function('return ' + cleaned))();
-    }
-  } catch (e) {
-    throw new Error(
-      `[parcel-resolver-storybook] Failed to load storybook globalsNameReferenceMap: ${(e as Error).message}`
-    );
-  }
-
-  for (const [specifier, globalName] of Object.entries(globalsNameReferenceMap)) {
-    const safeKey = specifier.replace(/[\/@]/g, '_');
-    const filePath = path.join(cacheDir, `${safeKey}.js`);
-
-    // Hardcode @storybook/global since its only export is `global`.
-    if (specifier === '@storybook/global') {
-      const code =
-        `// Generated by @parcel/resolver-storybook\n` +
-        `// Re-exports ${specifier} from globalThis.${globalName}\n` +
-        `const { global } = globalThis.${globalName};\n` +
-        `export { global };\n` +
-        `export default globalThis.${globalName};\n`;
-      fs.writeFileSync(filePath, code);
-      result[specifier] = { filePath, approach: 'B' };
-      continue;
-    }
-
-    // Approach B: enumerate named exports via require() and codegen
-    // a `const { ... } = globalThis.X; export { ... };` synthetic module.
-    let namedExports: string[] | null = null;
-    try {
-      const mod = _require(specifier);
-      if (mod && typeof mod === 'object') {
-        // Filter out 'default' and '__esModule' (CJS interop marker that some
-        // bundles attach as an enumerable own property).
-        namedExports = Object.keys(mod).filter(k => k !== 'default' && k !== '__esModule');
-      }
-    } catch {
-      namedExports = null;
-    }
-
-    if (namedExports && namedExports.length > 0) {
-      // Filter out reserved/invalid identifiers just in case
-      const valid = namedExports.filter(n => /^[A-Za-z_$][A-Za-z0-9_$]*$/.test(n));
-      const list = valid.join(', ');
-      const code =
-        `// Generated by @parcel/resolver-storybook\n` +
-        `// Re-exports ${specifier} from globalThis.${globalName}\n` +
-        `const { ${list} } = globalThis.${globalName};\n` +
-        `export { ${list} };\n` +
-        `export default globalThis.${globalName};\n`;
-      fs.writeFileSync(filePath, code);
-      result[specifier] = { filePath, approach: 'B' };
-    } else {
-      // Approach A fallback: emit CJS interop. Used when require() couldn't
-      // enumerate the module (e.g. ESM-only with no enumerable exports via require).
-      const code =
-        `// Generated by @parcel/resolver-storybook (Approach A fallback)\n` +
-        `// Could not enumerate named exports for ${specifier}; using CJS interop.\n` +
-        `module.exports = globalThis.${globalName};\n`;
-      fs.writeFileSync(filePath, code);
-      result[specifier] = { filePath, approach: 'A' };
-    }
-  }
-
-  return result;
-}
-
-// Generate cache files ONCE at module load (resolver-instance creation).
-// The runtime's setup() side-effect (in storybook/internal/preview/runtime)
-// populates the globals BEFORE any externalized cache file evaluates, so the
-// `const { x } = globalThis.__STORYBOOK_MODULE_*__` lines see the real exports.
-const STORYBOOK_EXTERNALS = buildExternalsCache();
-
 module.exports = new Resolver({
   async resolve({ dependency, options, specifier, pipeline, logger }) {
-    // Branch 1 - externalize storybook runtime specifiers (the patch-removal fix).
-    // Equivalent to webpack5's `externals: globalsNameReferenceMap` and Vite's
-    // alias-to-cache-file in storybook-external-globals-plugin.ts.
-    if (pipeline == null && Object.prototype.hasOwnProperty.call(STORYBOOK_EXTERNALS, specifier)) {
-      const entry = STORYBOOK_EXTERNALS[specifier];
-      return {
-        filePath: entry.filePath,
-        priority: 'sync',
-        pipeline: null,
-      };
-    }
-
-    // Branch 2 - React 17/18 interop shim. react-dom/client only exists from
-    // React 18 onwards; for React 17 we re-export react-dom so consumer code
-    // that does `import { createRoot } from 'react-dom/client'` still resolves.
+    // Workaround for interop issue
     if (specifier === "react-dom/client" && REACT_MAJOR_VERSION < 18) {
       return {
         filePath: __dirname + "/react.js",
@@ -173,13 +20,8 @@ module.exports = new Resolver({
       };
     }
 
-    // Branch 3 - the `story:` pipeline. The specifier is a base64-encoded glob
-    // (relative to the importing file's dir). We expand the glob and emit a
-    // synthetic ESM module whose default export maps cwd-relative paths to
-    // dynamic-import factories. Doing this in a resolver (rather than a
-    // transformer) lets Parcel's watcher invalidate when matching files appear
-    // or disappear -- see `invalidateOnFileCreate` below. Storybook's importFn
-    // expects keys relative to process.cwd().
+    // Resolve story entry globs. Storybook expects an object with relative paths from the process cwd as keys.
+    // We do this in a resolver so that it invalidates the watcher when new stories are created.
     if (pipeline === 'story') {
       let sourceFile = dependency.resolveFrom ?? dependency.sourcePath!;
       let normalized = normalizeSeparators(path.resolve(path.dirname(sourceFile), atob(specifier)));
@@ -200,7 +42,7 @@ module.exports = new Resolver({
           dir,
           'stories.js'
         ),
-        code: `export default {\n${results.join('\n')}\n};\n`,
+        code: `module.exports = {\n${results.join('\n')}\n};\n`,
         invalidateOnFileCreate: [
           {glob: normalized}
         ],
