+++ packages/dev/parcel-transformer-mdx-storybook/StorybookMDXTransformer.mjs
@@ -1,5 +1,3 @@
-// Compiles MDX docs pages to JSX using @mdx-js/mdx with the @mdx-js/react
-// provider so Storybook's docs renderer can wrap them in MDXProvider.
 import {compile} from '@mdx-js/mdx';
 import {Transformer} from '@parcel/plugin';
