+++ packages/dev/parcel-transformer-storybook/package.json
@@ -16,7 +16,7 @@
     "@parcel/plugin": "^2.16.3",
     "@parcel/source-map": "^2.1.1",
     "react-docgen-typescript": "^2.2.2",
-    "storybook": "^10.0.0",
+    "storybook": "^9.0.18",
     "typescript": "^5.5.0"
   },
   "scripts": {
