+++ packages/dev/storybook-builder-parcel/gen-iframe-modern.js
@@ -1,27 +1,47 @@
-// Template extracted verbatim from upstream Vite's
-// code/builders/builder-vite/input/iframe.html (Task 2.2 of the upstream-
-// alignment plan). Placeholder strings are filled in below via String.replace.
+const { normalizeStories } = require("storybook/internal/common");
 
-import { normalizeStories } from "storybook/internal/common";
-import { readFileSync } from "node:fs";
-import { fileURLToPath } from "node:url";
+const TEMPLATE = `<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <title><!-- [TITLE HERE] --></title>
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <script>
+      window.CONFIG_TYPE = '[CONFIG_TYPE HERE]';
+      window.LOGLEVEL = '[LOGLEVEL HERE]';
+      window.FRAMEWORK_OPTIONS = '[FRAMEWORK_OPTIONS HERE]';
+      window.CHANNEL_OPTIONS = '[CHANNEL_OPTIONS HERE]';
+      window.FEATURES = '[FEATURES HERE]';
+      window.STORIES = '[STORIES HERE]';
+      window.DOCS_OPTIONS = '[DOCS_OPTIONS HERE]';
+      window.SERVER_CHANNEL_URL = '[SERVER_CHANNEL_URL HERE]';
+    </script>
+    <!-- [HEAD HTML SNIPPET HERE] -->
+  </head>
+  <body>
+    <!-- [BODY HTML SNIPPET HERE] -->
+    <div id="storybook-root"></div>
+    <div id="storybook-docs"></div>
+    <script type="module" src="preview.js"></script>
+  </body>
+</html>
+`;
 
-const TEMPLATE_PATH = fileURLToPath(new URL("./templates/iframe.html", import.meta.url));
-
-export async function generateIframeModern(options) {
-  const TEMPLATE = readFileSync(TEMPLATE_PATH, "utf8");
+module.exports.generateIframeModern = async function generateIframeModern(options) {
   const { configType, features, presets, serverChannelUrl, title } = options;
   const frameworkOptions = await presets.apply("frameworkOptions");
   const headHtmlSnippet = await presets.apply("previewHead");
   const bodyHtmlSnippet = await presets.apply("previewBody");
   const logLevel = await presets.apply("logLevel", undefined);
   const docsOptions = await presets.apply("docs");
-  const tagsOptions = await presets.apply("tags", {});
 
   const coreOptions = await presets.apply("core");
   const stories = normalizeStories(
     await options.presets.apply("stories", [], options),
-    { configDir: options.configDir, workingDir: process.cwd() }
+    {
+      configDir: options.configDir,
+      workingDir: process.cwd(),
+    }
   ).map((specifier) => ({
     ...specifier,
     importPathMatcher: specifier.importPathMatcher.source,
@@ -34,18 +54,16 @@ export async function generateIframeModern(options) {
     .replace(`'[FRAMEWORK_OPTIONS HERE]'`, JSON.stringify(frameworkOptions))
     .replace(
       `'[CHANNEL_OPTIONS HERE]'`,
-      JSON.stringify(coreOptions?.channelOptions ?? {})
+      JSON.stringify(
+        coreOptions && coreOptions.channelOptions
+          ? coreOptions.channelOptions
+          : {}
+      )
     )
     .replace(`'[FEATURES HERE]'`, JSON.stringify(features || {}))
     .replace(`'[STORIES HERE]'`, JSON.stringify(stories || {}))
     .replace(`'[DOCS_OPTIONS HERE]'`, JSON.stringify(docsOptions || {}))
-    .replace(`'[TAGS_OPTIONS HERE]'`, JSON.stringify(tagsOptions || {}))
     .replace(`'[SERVER_CHANNEL_URL HERE]'`, JSON.stringify(serverChannelUrl))
-    // OTHER_GLOBALS is a Vite-only injection point used by their externalize
-    // plugin to splat extra globals into the inline script. We don't use it;
-    // the placeholder must be replaced with a syntactically-valid expression
-    // (we use empty string `''`) -- leaving bare `()` would crash the parser.
-    .replace(`'[OTHER_GLOBALS HERE]'`, `''`)
     .replace("<!-- [HEAD HTML SNIPPET HERE] -->", headHtmlSnippet || "")
     .replace("<!-- [BODY HTML SNIPPET HERE] -->", bodyHtmlSnippet || "");
-}
+};
