+++ packages/dev/storybook-builder-parcel/package.json
@@ -2,18 +2,12 @@
   "name": "storybook-builder-parcel",
   "version": "0.0.1",
   "private": true,
-  "type": "module",
-  "main": "./preset.mjs",
-  "exports": {
-    ".": "./preset.mjs",
-    "./preset": "./preset.mjs"
-  },
+  "main": "preset.js",
   "dependencies": {
     "@parcel/core": "^2.16.3",
     "@parcel/reporter-cli": "^2.16.3",
-    "@parcel/utils": "^2.16.3",
     "http-proxy-middleware": "^2.0.6",
-    "storybook": "^10.0.0"
+    "storybook": "^9.0.18"
   },
   "peerDependencies": {
     "react": "*"
