--- packages/dev/storybook-builder-parcel/templates/iframe.html
@@ -1,39 +1,0 @@
-<!--
-  Mirrors upstream code/builders/builder-vite/input/iframe.html.
-  Placeholder strings (e.g. '[CONFIG_TYPE HERE]', '[FEATURES HERE]') are
-  substituted at build time by gen-iframe-modern.mjs via String.replace.
--->
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title><!-- [TITLE HERE] --></title>
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <script>
-      window.CONFIG_TYPE = '[CONFIG_TYPE HERE]';
-      window.LOGLEVEL = '[LOGLEVEL HERE]';
-      window.FRAMEWORK_OPTIONS = '[FRAMEWORK_OPTIONS HERE]';
-      window.CHANNEL_OPTIONS = '[CHANNEL_OPTIONS HERE]';
-      window.FEATURES = '[FEATURES HERE]';
-      window.STORIES = '[STORIES HERE]';
-      window.DOCS_OPTIONS = '[DOCS_OPTIONS HERE]';
-      window.TAGS_OPTIONS = '[TAGS_OPTIONS HERE]';
-      window.SERVER_CHANNEL_URL = '[SERVER_CHANNEL_URL HERE]';
-
-      ('[OTHER_GLOBALS HERE]');
-
-      // Preserves Storybook code paths that path-sniff `module && module.hot`
-      // (treated as "no HMR available") and code that expects a `global`
-      // reference (legacy node-style). Matches upstream Vite's iframe.html.
-      window.module = undefined;
-      window.global = window;
-    </script>
-    <!-- [HEAD HTML SNIPPET HERE] -->
-  </head>
-  <body>
-    <!-- [BODY HTML SNIPPET HERE] -->
-    <div id="storybook-root"></div>
-    <div id="storybook-docs"></div>
-    <script type="module" src="preview-main.js"></script>
-  </body>
-</html>
