+++ packages/dev/storybook-react-parcel/package.json
@@ -2,25 +2,20 @@
   "name": "storybook-react-parcel",
   "version": "0.0.1",
   "private": true,
-  "type": "module",
-  "main": "./preset.mjs",
-  "exports": {
-    ".": "./preset.mjs",
-    "./preset": "./preset.mjs"
-  },
+  "main": "preset.js",
   "dependencies": {
-    "@storybook/react": "^10.0.0",
+    "@storybook/react": "^9.0.18",
     "storybook-builder-parcel": ">=0.0.1"
   },
   "devDependencies": {
     "react": "^16.8.0 || ^17.0.0-rc.1 || ^18.0.0 || ^19.0.0",
     "react-dom": "^16.8.0 || ^17.0.0-rc.1 || ^18.0.0 || ^19.0.0",
-    "storybook": "^10.0.0"
+    "storybook": "^9.0.18"
   },
   "peerDependencies": {
     "react": "^16.8.0 || ^17.0.0-rc.1 || ^18.0.0 || ^19.0.0",
     "react-dom": "^16.8.0 || ^17.0.0-rc.1 || ^18.0.0 || ^19.0.0",
-    "storybook": "^10.0.0"
+    "storybook": "^9.0.18"
   },
   "scripts": {
     "build": "true",
