+++ packages/dev/storybook-react-parcel/preset.js
@@ -1,0 +1,4 @@
+module.exports.core = {
+	builder: "storybook-builder-parcel",
+	renderer: "@storybook/react",
+};
